[BRK] move errors to src/errors, replace csv parser errors with new errors, update unit tests for generic csv parser
