+++ tsconfig.json
@@ -52,7 +52,8 @@
     "moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */,
     "baseUrl": "src" /* Base directory to resolve non-absolute module names. */,
     "paths": {
-      "@/helper": ["helper"]
+      "@/helper": ["helper"],
+      "@/errors": ["errors"]
     } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
     // "rootDirs": [],                              /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                             /* List of folders to include type definitions from. */
